+ exception handling v REST architekture

+++ eshop-web/src/main/java/com/example/eshopweb/service/ItemService.java
@@ -2,6 +2,7 @@ package com.example.eshopweb.service;
 
 import com.example.eshopweb.dto.entity.ItemDto;
 import com.example.eshopweb.entity.Item;
+import com.example.eshopweb.exception.NotFoundException;
 import com.example.eshopweb.mapper.ItemMapper;
 import com.example.eshopweb.repository.ItemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -38,6 +39,9 @@ public class ItemService {
     }
 
     public void deleteById(int id) {
+        if(!itemRepository.existsById(id)) {
+            throw new NotFoundException("Item with id: " + id + " does not exist!");
+        }
         itemRepository.deleteById(id);
     }
